refactor: :recycle: adds hooks directory and moves hooks there
renames custom hook modules to have a .hook.js extension cuz it looks cool 😎
