+++ frontend/src/components/UI/WizardSteps/StepCreateBudgetLines/StepCreateBudgetLines.jsx
@@ -13,7 +13,7 @@ import suite from "./suite";
 import { convertCodeForDisplay } from "../../../../helpers/utils";
 import ConfirmationModal from "../../Modals/ConfirmationModal";
 import { useUpdateBudgetLineItemMutation, useAddBudgetLineItemMutation } from "../../../../api/opsAPI";
-import useAlert from "../../../../helpers/use-alert";
+import useAlert from "../../../../hooks/use-alert.hooks";
 
 /**
  * Renders the Create Budget Lines component with React context.
