+++ frontend/src/pages/agreements/EditAgreement.js
@@ -6,7 +6,7 @@ import CreateEditAgreement from "./CreateEditAgreement";
 import { useGetAgreementByIdQuery } from "../../api/opsAPI";
 import { getUser } from "../../api/getUser";
 import SimpleAlert from "../../components/UI/Alert/SimpleAlert";
-import { useIsAgreementEditable, useIsUserAllowedToEditAgreement } from "../../helpers/agreement-hooks";
+import { useIsAgreementEditable, useIsUserAllowedToEditAgreement } from "../../hooks/agreement.hooks";
 
 const EditAgreement = () => {
     const urlPathParams = useParams();
